HT2 data-access folder renamed controller's types prettified
