+++ src/utils/findInstance.ts
@@ -1,7 +1,7 @@
 /**
- *
- * @param array of instances where utility will find the wanted one for you
- * @param fieldName utility will search an instance by this param
+ * Find object in an Array of Objects by a field.
+ * @param array of instances where utility will find the wanted one for you.
+ * @param fieldName utility will search an instance by this field.
  * @param wantedFieldValue expected value of a `feildName`
  * @returns array of 2 elements, `index` of an instance in array, and a found `instance` itself
  */
